+++ source/djWebFilterHolder.pas
@@ -83,7 +83,7 @@ type
      *
      * @param Key init parameter name (case-sensitive)
      * @param Value init parameter value
-     * @throws EWebComponentException if the key is already set
+     * @throws EDarajaConfigException if the key is already set
      *}
     procedure SetInitParameter(const Key: string; const Value: string);